board 테이블 insert, select 완료

+++ src/main/java/com/todo/spring/domain/repository/ProductRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.todo.spring.domain.entity.ProductManagement;
 
-public interface ProductRepository extends JpaRepository<ProductManagement, Long> {
+public interface ProductRepository extends JpaRepository<ProductManagement, Integer> {
 
 }
